requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Banco de dados assíncrono
sqlalchemy==2.0.36
aiosqlite==0.20.0

# Autenticação e segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Leitura de variáveis de ambiente (.env)
python-dotenv==1.0.1

# Validação e tipagem
pydantic[email]==2.9.2
